lint: remove the var decl check
Removes the var decl check which often fires when the type is required
for documentation purposes. For example, when an enumerated list of
pre-defined values of a certain type is created, it may be desirable to
group those values under the type itself in the docs. If the type is
infered, godoc will not group them, but if the type is manually listed
they will be placed in the correct location.

Since this lint is not an important stylistic consideration and
frequently produces false positives, removing it is consistent with the
scope of lints in this project.

Fixes golang/lint#429

Change-Id: I2afd928eae0bb11b39886c045556cda0771e3af5
Reviewed-on: https://go-review.googlesource.com/c/lint/+/191317
Run-TryBot: Sam Whited <sam@samwhited.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Andrew Bonventre <andybons@golang.org>